requirements.txt
flask
flask-socketio
python-chess
requests
torch
